# [Bugfix] hotfix for gptq-marlin non-contiguous error