@DefTruth 2025-03-25
Close, since https://github.com/vllm-project/vllm/pull/14658 has been merged.